+++ perl/FSM/Scheduler/ISF.pm
@@ -3,6 +3,7 @@ package FSM::Scheduler::ISF;
 use v5.20;
 use strict;
 use warnings;
+use Carp qw(confess);
 use feature qw(signatures postderef);
 no warnings 'experimental::signatures';
 
@@ -11,7 +12,8 @@ use FSM::Scheduler::ISF::LoweringIR;
 use FSM::Scheduler::ISF::Emitter::FSM;
 use FSM::Scheduler::ISF::Emitter::JSON;
 
-sub new($class, %args) {
+sub new($class, @constructor_args) {
+    my %args = _validate_constructor_args($class, @constructor_args);
     my $debug = $args{debug} // 0;
     fsm_trace_enter('Initialize ISF scheduler', 2);
 
@@ -24,6 +26,20 @@ sub new($class, %args) {
 
     fsm_trace_exit('ISF scheduler initialized', 2);
     return $self;
+}
+
+sub _validate_constructor_args($class, @args) {
+    confess "$class->new expects an even-length option/value list\n"
+        if @args % 2;
+
+    my %options = @args;
+    my %allowed = map { $_ => 1 } qw(debug);
+    for my $name (sort keys %options) {
+        confess "$class->new unsupported option '$name'; supported option: debug\n"
+            unless $allowed{$name};
+    }
+
+    return %options;
 }
 
 sub lower($self, $actor) {